Use a constant for duration
The rationale for this is that the time to record/play buffers takes an order of
magnitude (milliseconds vs micro/nanoseconds) longer than the time to do the
other stuff, so it's safe to base the delay off that.

Also add a fudge factor, since I think it's better to not make it all the way to
the end of a slope than to start disappearing and pulsing.