Fix divide by zero bug
If the user is able to generate two successive formants with the same F2, slope
is set to a division by zero, which becomes negative infinity. When slope is
used in the calculation for the next y point, it implicitely widens the entire
expression and gives it the value of negative infinity. This values is then
implicitely narrowed to a short and becomes zero.

The result of this is the tracer would temporarily bug out to a y values of zero
and slide back down.

To fix this, narrow the slope calculation before adding it to the offset